requirements.txt
aiohttp>=3.9,<4.0
websockets>=12.0,<16.0
python-dotenv>=1.0,<2.0
